# perf: Reduce external repo pointer allocations