--- a/cmd/frontend/authz/common.go
+++ b/cmd/frontend/authz/common.go
@@ -8,11 +8,11 @@ import (
 func ToRepos(src []*types.Repo) (dst map[Repo]struct{}) {
 	dst = make(map[Repo]struct{})
 	for _, r := range src {
-		rp := Repo{ID: r.ID, RepoName: r.Name}
-		if r.ExternalRepo != nil {
-			rp.ExternalRepoSpec = *r.ExternalRepo
-		}
-		dst[rp] = struct{}{}
+		dst[Repo{
+			ID:               r.ID,
+			RepoName:         r.Name,
+			ExternalRepoSpec: r.ExternalRepo,
+		}] = struct{}{}
 	}
 	return dst
 }
@@ -20,7 +20,7 @@ func ToRepos(src []*types.Repo) (dst map[Repo]struct{}) {
 func GetCodeHostRepos(c *extsvc.CodeHost, repos map[Repo]struct{}) (mine map[Repo]struct{}, others map[Repo]struct{}) {
 	mine, others = make(map[Repo]struct{}), make(map[Repo]struct{})
 	for repo := range repos {
-		if extsvc.IsHostOf(c, &repo.ExternalRepoSpec) {
+		if extsvc.IsHostOf(c, repo.ExternalRepoSpec) {
 			mine[repo] = struct{}{}
 		} else {
 			others[repo] = struct{}{}