--- a/cmd/frontend/backend/repos_mock.go
+++ b/cmd/frontend/backend/repos_mock.go
@@ -19,10 +19,10 @@ type MockRepos struct {
 	GetByName                 func(v0 context.Context, name api.RepoName) (*types.Repo, error)
 	AddGitHubDotComRepository func(name api.RepoName) error
 	List                      func(v0 context.Context, v1 db.ReposListOptions) ([]*types.Repo, error)
-	GetCommit                 func(v0 context.Context, repo types.RepoIdentifier, commitID api.CommitID) (*git.Commit, error)
-	ResolveRev                func(v0 context.Context, repo types.RepoIdentifier, rev string) (api.CommitID, error)
-	GetInventory              func(v0 context.Context, repo types.RepoIdentifier, commitID api.CommitID) (*inventory.Inventory, error)
-	GetInventoryUncached      func(ctx context.Context, repo types.RepoIdentifier, commitID api.CommitID) (*inventory.Inventory, error)
+	GetCommit                 func(v0 context.Context, repo *types.Repo, commitID api.CommitID) (*git.Commit, error)
+	ResolveRev                func(v0 context.Context, repo *types.Repo, rev string) (api.CommitID, error)
+	GetInventory              func(v0 context.Context, repo *types.Repo, commitID api.CommitID) (*inventory.Inventory, error)
+	GetInventoryUncached      func(ctx context.Context, repo *types.Repo, commitID api.CommitID) (*inventory.Inventory, error)
 }
 
 var errRepoNotFound = &errcode.Mock{
@@ -85,7 +85,7 @@ func (s *MockRepos) MockList(t *testing.T, wantRepos ...api.RepoName) (called *b
 func (s *MockRepos) MockResolveRev_NoCheck(t *testing.T, commitID api.CommitID) (called *bool) {
 	var once sync.Once
 	called = new(bool)
-	s.ResolveRev = func(ctx context.Context, repo types.RepoIdentifier, rev string) (api.CommitID, error) {
+	s.ResolveRev = func(ctx context.Context, repo *types.Repo, rev string) (api.CommitID, error) {
 		once.Do(func() {
 			*called = true
 		})
@@ -96,22 +96,22 @@ func (s *MockRepos) MockResolveRev_NoCheck(t *testing.T, commitID api.CommitID) 
 
 func (s *MockRepos) MockResolveRev_NotFound(t *testing.T, wantRepo api.RepoID, wantRev string) (called *bool) {
 	called = new(bool)
-	s.ResolveRev = func(ctx context.Context, repo types.RepoIdentifier, rev string) (api.CommitID, error) {
+	s.ResolveRev = func(ctx context.Context, repo *types.Repo, rev string) (api.CommitID, error) {
 		*called = true
-		if repo.RepoID() != wantRepo {
-			t.Errorf("got repo %v, want %v", repo.RepoID(), wantRepo)
+		if repo.ID != wantRepo {
+			t.Errorf("got repo %v, want %v", repo.ID, wantRepo)
 		}
 		if rev != wantRev {
 			t.Errorf("got rev %v, want %v", rev, wantRev)
 		}
-		return "", &git.RevisionNotFoundError{Repo: repo.RepoName(), Spec: rev}
+		return "", &git.RevisionNotFoundError{Repo: repo.Name, Spec: rev}
 	}
 	return
 }
 
 func (s *MockRepos) MockGetCommit_Return_NoCheck(t *testing.T, commit *git.Commit) (called *bool) {
 	called = new(bool)
-	s.GetCommit = func(ctx context.Context, repo types.RepoIdentifier, commitID api.CommitID) (*git.Commit, error) {
+	s.GetCommit = func(ctx context.Context, repo *types.Repo, commitID api.CommitID) (*git.Commit, error) {
 		*called = true
 		return commit, nil
 	}