--- a/cmd/frontend/backend/repos_vcs.go
+++ b/cmd/frontend/backend/repos_vcs.go
@@ -24,36 +24,36 @@ import (
 // knowing the remote URL is necessary to perform any operations (from method calls on the return
 // value), those operations will fail. This occurs when the repository isn't cloned on gitserver or
 // when an update is needed (eg in ResolveRevision).
-func CachedGitRepo(ctx context.Context, repo types.RepoIdentifier) (*gitserver.Repo, error) {
-	r, err := quickGitserverRepo(ctx, repo.RepoName(), repo.ExternalRepoSpec().ServiceType)
+func CachedGitRepo(ctx context.Context, repo *types.Repo) (*gitserver.Repo, error) {
+	r, err := quickGitserverRepo(ctx, repo.Name, repo.ExternalRepo.ServiceType)
 	if err != nil {
 		return nil, err
 	}
 	if r != nil {
 		return r, nil
 	}
-	return &gitserver.Repo{Name: repo.RepoName()}, nil
+	return &gitserver.Repo{Name: repo.Name}, nil
 }
 
 // GitRepo returns a handle to the Git repository with the up-to-date (as of the time of this call)
 // remote URL. See CachedGitRepo for when this is necessary vs. unnecessary.
-func GitRepo(ctx context.Context, repo types.RepoIdentifier) (gitserver.Repo, error) {
-	gitserverRepo, err := quickGitserverRepo(ctx, repo.RepoName(), repo.ExternalRepoSpec().ServiceType)
+func GitRepo(ctx context.Context, repo *types.Repo) (gitserver.Repo, error) {
+	gitserverRepo, err := quickGitserverRepo(ctx, repo.Name, repo.ExternalRepo.ServiceType)
 	if err != nil {
-		return gitserver.Repo{Name: repo.RepoName()}, err
+		return gitserver.Repo{Name: repo.Name}, err
 	}
 	if gitserverRepo != nil {
 		return *gitserverRepo, nil
 	}
 
 	result, err := repoupdater.DefaultClient.RepoLookup(ctx, protocol.RepoLookupArgs{
-		Repo: repo.RepoName(),
+		Repo: repo.Name,
 	})
 	if err != nil {
-		return gitserver.Repo{Name: repo.RepoName()}, err
+		return gitserver.Repo{Name: repo.Name}, err
 	}
 	if result.Repo == nil {
-		return gitserver.Repo{Name: repo.RepoName()}, repoupdater.ErrNotFound
+		return gitserver.Repo{Name: repo.Name}, repoupdater.ErrNotFound
 	}
 	return gitserver.Repo{Name: result.Repo.Name, URL: result.Repo.VCS.URL}, nil
 }
@@ -141,12 +141,12 @@ func hasGitLabDotComToken(ctx context.Context) (bool, error) {
 // * Empty repository: git.RevisionNotFoundError
 // * The user does not have permission: errcode.IsNotFound
 // * Other unexpected errors.
-func (s *repos) ResolveRev(ctx context.Context, repo types.RepoIdentifier, rev string) (commitID api.CommitID, err error) {
+func (s *repos) ResolveRev(ctx context.Context, repo *types.Repo, rev string) (commitID api.CommitID, err error) {
 	if Mocks.Repos.ResolveRev != nil {
 		return Mocks.Repos.ResolveRev(ctx, repo, rev)
 	}
 
-	ctx, done := trace(ctx, "Repos", "ResolveRev", map[string]interface{}{"repo": repo.RepoName(), "rev": rev}, &err)
+	ctx, done := trace(ctx, "Repos", "ResolveRev", map[string]interface{}{"repo": repo.Name, "rev": rev}, &err)
 	defer done()
 
 	// We start out by using a CachedGitRepo which doesn't have a remote URL.
@@ -168,15 +168,15 @@ func (s *repos) ResolveRev(ctx context.Context, repo types.RepoIdentifier, rev s
 	return git.ResolveRevision(ctx, *gitserverRepo, remoteURLFunc, rev, nil)
 }
 
-func (s *repos) GetCommit(ctx context.Context, repo types.RepoIdentifier, commitID api.CommitID) (res *git.Commit, err error) {
+func (s *repos) GetCommit(ctx context.Context, repo *types.Repo, commitID api.CommitID) (res *git.Commit, err error) {
 	if Mocks.Repos.GetCommit != nil {
 		return Mocks.Repos.GetCommit(ctx, repo, commitID)
 	}
 
-	ctx, done := trace(ctx, "Repos", "GetCommit", map[string]interface{}{"repo": repo.RepoName(), "commitID": commitID}, &err)
+	ctx, done := trace(ctx, "Repos", "GetCommit", map[string]interface{}{"repo": repo.Name, "commitID": commitID}, &err)
 	defer done()
 
-	log15.Debug("svc.local.repos.GetCommit", "repo", repo.RepoName(), "commitID", commitID)
+	log15.Debug("svc.local.repos.GetCommit", "repo", repo.Name, "commitID", commitID)
 
 	if !git.IsAbsoluteRevision(string(commitID)) {
 		return nil, errors.Errorf("non-absolute CommitID for Repos.GetCommit: %v", commitID)
@@ -206,7 +206,7 @@ func isIgnorableRepoUpdaterError(err error) bool {
 	return err == repoupdater.ErrNotFound || err == repoupdater.ErrUnauthorized || err == repoupdater.ErrTemporarilyUnavailable
 }
 
-func maybeLogRepoUpdaterError(repo types.RepoIdentifier, err error) {
+func maybeLogRepoUpdaterError(repo *types.Repo, err error) {
 	var msg string
 	switch c := errors.Cause(err); c {
 	case repoupdater.ErrNotFound:
@@ -217,6 +217,6 @@ func maybeLogRepoUpdaterError(repo types.RepoIdentifier, err error) {
 		msg = "Repository host was temporarily unavailable while retrieving repository information."
 	}
 	if msg != "" {
-		log15.Warn(msg+" Consult repo-updater logs for more information.", "repo", repo.RepoName())
+		log15.Warn(msg+" Consult repo-updater logs for more information.", "repo", repo.Name)
 	}
 }