--- a/cmd/frontend/graphqlbackend/repository_external.go
+++ b/cmd/frontend/graphqlbackend/repository_external.go
@@ -11,10 +11,6 @@ import (
 )
 
 func (r *repositoryResolver) ExternalRepository() *externalRepositoryResolver {
-	if r.repo.ExternalRepo == nil {
-		return nil
-	}
-
 	return &externalRepositoryResolver{repository: r}
 }