--- a/cmd/frontend/graphqlbackend/repository_mirror.go
+++ b/cmd/frontend/graphqlbackend/repository_mirror.go
@@ -204,7 +204,7 @@ func (r *schemaResolver) CheckMirrorRepositoryConnection(ctx context.Context, ar
 		}
 	case args.Name != nil:
 		// GitRepo will use just the name to look up the repository from repo-updater.
-		repo = types.NewRepoWithIDs(0, api.RepoName(*args.Name), nil)
+		repo = &types.Repo{RepoIDs: types.RepoIDs{Name: api.RepoName(*args.Name)}}
 	}
 
 	gitserverRepo, err := backend.GitRepo(ctx, repo)